Add a Task State Segment
Intel requires this structure to be set up to change rings - something that
is needed for context switching to userspace. Most of its contents are
irrelevant, with the exception of the kernel stack; that is put into place
when a syscall occurs. The kernel stack should be allocated on thread-launch
on a per-thread basis.