Add initrd loading and basic info parsing
The initrd directory holds all files that are to be included into the initrd
image. A tarball of it is then copied right next to the kernel image and
loaded by GRUB as a multiboot2 module. The kernel then locates the initrd,
maps it into memory and adjusts the physical address to the new virtual one.
A utility function converts the base 8 ASCII-written size into a size_t.
Reading files is as simple as obtaining a pointer to them by adding 0x200
to their header address.