Add kernel worker threads and functionality to pass arguments to threads
Kernel worker threads are a dependency for ACPICA. They are passed exactly one
argument of type void *. User threads are passed envc, envp, argc and argv for
now. In the future, libc should take care of envc and envp, so that main only
gets argc and argv passed to it.