Strip toolchain building as prepping for dockerized build environment
Building gcc, binutils and GRUB (to name the worst offenders)
consistently on varying hosts proved difficult, so moving to a
standardized environment provided by docker containers solves this.
That way, the toolchain is built on my local machine and distributed
via the Docker Hub, thus removing the need to compile any software
but QEMU, which unfortunately doesn't simply run in a container.
For now, support for local toolchains is completely stripped from
the project, but supplying PKGBUILDs and .debs is under consideration
to allow for container-less environments again.